+++ packages/atomic-angular/projects/atomic-angular/CHANGELOG.md
@@ -1,3 +1,18 @@
+## 3.10.0 (2026-01-07)
+
+* [Version Bump][skip ci]: ui-kit publish ([b2a83c4](https://github.com/coveo/ui-kit/commits/b2a83c4))
+* Add Node.js 24 (Krypton LTS) support and update CI (#6756) ([1ee69fa](https://github.com/coveo/ui-kit/commits/1ee69fa)), closes [#6756](https://github.com/coveo/ui-kit/issues/6756)
+* feat(atomic-angular): support angular 20 (#6813) ([caa0b39](https://github.com/coveo/ui-kit/commits/caa0b39)), closes [#6813](https://github.com/coveo/ui-kit/issues/6813)
+
+
+
+## 3.9.0 (2026-01-06)
+
+* Add Node.js 24 (Krypton LTS) support and update CI (#6756) ([1ee69fa](https://github.com/coveo/ui-kit/commits/1ee69fa)), closes [#6756](https://github.com/coveo/ui-kit/issues/6756)
+* feat(atomic-angular): support angular 20 (#6813) ([caa0b39](https://github.com/coveo/ui-kit/commits/caa0b39)), closes [#6813](https://github.com/coveo/ui-kit/issues/6813)
+
+
+
 ## 3.8.0 (2025-11-12)
 
 * chore: remove stencil-generated build artifacts from git tracking (#6500) ([e9f418f](https://github.com/coveo/ui-kit/commits/e9f418f)), closes [#6500](https://github.com/coveo/ui-kit/issues/6500)